parted global cache only create cache on nodes with files.
The cache size created on the node depends on the size of data in the
node. Then we need to define a mapping function accordingly.

Now the problem is that the threads run on the nodes are overloaded
but the threads on other nodes are busy waiting, which keeps locking the
queues of other nodes and checking them. Maybe using mutex is a better
option.