Enhance the flush thread.
Now the flush thread is enabled by default. The main change is to
distinguish requests initiated by applications and those by the flush
thread. The requests initiated by the flush thread should be processed
in the background.

The main idea is to provide IO threads enough requests to process, so
when it slows down for garbage collection, other threads still have
requests to process.
The other benefit is that synchronous writes to the page cache can
still trigger parallel writes and achieve the maximal performance.