Add low-latency flag in io_request.
It is useful for global_cached_io. An application IO may cause extra IO
in global_cached_io. For example, when the page cache evicts a dirty
page, we need to write back the dirty page. The writeback should be
completed with low latency, so that the I/O thread won't buffer it.
Otherwise, there may be a dead lock. global_cached_io is waiting for
requests to complete, but the I/O thread buffers the notification of the
completion of the requests.

The deadlock also exists in remote_disk_access, but it's unlikely to
occur if we always send as many parallel IO as allowed by the system.
However, this is very likely to happen in global_cached_io with sync
write.