Limiting the number of flushes.
As a result, we need to increase the number of flush requests from the
cache. Three changes are made:
Flush dirty pages in notify_completion.
add dirty cells back in the queue if they have more dirty pages for
flushing;
flush dirty cells no matter whehter they are in the queue.