[SAFS]: redesign the naming of partition files in local filesystems.
With the new design, the library can recognize the order of partitions
files, regardless the order of the root paths of the RAID in the configuration
file.

The order of the root paths in the configuration file may change due to
various reasons, so it's better not to rely on it.