[SAFS][simp-orig-1]: Attempt to remove partial original reqs in global_cached_io.
Currently, global_cached_io needs to allocate an I/O request to keep the
info of the original user request. In addition, it may need to allocate
I/O requests to represent part of the original user request.

The patch attempts to remove the partial original requests to simplify
the code.