[SAFS]: global_cached_io accepts an arbitrary number of reqs.
global_cached_io stops sending requests to the underlying IO when it
can't reference a page. It queues the user's requests to an internal
queue. It resumes issuing requests when the application waits for
requests to complete.

Dirty page flusher writes too many pages too aggressively. I need to
change it.

[SAFS]: fix a bug part_global_cached_io.