[SAFS][adv-comp-5]: fix a bug in global_cached_io.
We should postpone counting the number of completed async requests.

There is still a problem left. In the case of part_global_cached_io, the
requests generated in user compute will be issued by the
global_cached_io that execute the user compute. So it's hard to check if
all requests have been completed. I need to re-implement
part_global_cached_io for user compute.