[Graph]: fix a potential bug in the graph engine.
The condition of determining the end of an iteration in the original
implementation to test the number of pending I/O requests in the IO
instance. However, this may be insufficient because user compute may
haven't issued all requests yet. Although it's unclear whether the bug
can be triggered, it's better to count the number of incomplete vertices
in the graph engine and then determine the condition of the end of an
iteration.