[SAFS][adv-comp-20]: allow user compute to issue reqs many times.
The previous implementation has a temporary solution for user compute to
issue requests. User compute can issue requests only once. This commit
is to fix the problem and allow user compute to issue as many requests
as possible.

Note: this commit seems to have some bugs about the reference of a user
compute. The next commit fixes the problem.