[Graph]: create vertex_index optimized for NUMA.
The new vertex_index splits the graph into the same number of partitions
as the number of threads specified by the user. Each thread gets a
partition.

When we allocate memory for the partition of vertex_index on the same
node as the thread runs, we can see significant performance improvement.

Only BFS and SSSP start to use the new vertex_index.