[Graph]: every vertex computes local scan all by itself.
This commit disables the optimization that only the largest vertex in
a triangle counts edges. This optimization may have a potential bug that
the largest vertex in a triangle is ignored because its potential max
local scan is smaller than the current max scan.

The simple solution to fix the bug is that every vertex computes local
scan itself.