[Graph]: fix a bug in load balancer.
The previous implementation allows vertices stolen from other threads to
be stolen again. It's possible that a vertex stolen by another thread
ends up to be processed in its owner thread. It will mess up counting
the number of stolen vertices and returned vertices.

This fix simplifies the process of stealing vertices from other threads.
Each time a thread only steals the number of vertices that it can
process.