[Graph]: fix a memory leak.
The graph engine expects a vertex to request complete vertex
information from the SSDs at the first time. If a vertex requests a
partial vertex at the first time, the original vertex_compute is created
but not sent to the filesystem and thus can't be deleted. We need to
delete them explicitly.