[Graph]: fix a bug of issuing vertex_compute to SAFS.
Originally, the graph engine didn't hold a reference count on
vertex_compute. Now each worker thread maintains a set of vertex_compute
for active vertices and hold a reference count on vertex_compute. We
need to change the way of determining whether a vertex_compute has been
issued to SAFS.