[Graph]: issue all vertex_compute to SAFS.
It's possible that directed_vertex_compute is not issued to SAFS and is
maintained by the graph engine. It happens when the graph algorithm only
requests partial vertices. This commit makes sure all vertex_compute are
maintained by SAFS, so SAFS can maintain the number of pending I/O
requests it wants.