[Matrix]: change task scheduling in SpMM.
We use a global queue to maintain all tasks in SpMM and each worker
thread fetches one task at a time from the queue. As such, when we need
to write the portions of the output matrix to SSDs, the portions written
to SSDs in a short period of time are logically close to each other.
Given the distribution of the portions on SSDs, the data written to SSDs
are evenly distributed among all SSDs.