[Matrix]: reduce NUMA range size.
This is a temp solution. The range size determines the block size used
in a chunked_raw_array. When the range size is large, we'll have to keep
many rows in the same memory chunk. However, if a tall dense matrix has
many columns, a memory chunk can't fit all rows in a range.